+++ Hytale/Features/ActualFeatures/NoFall.cpp
@@ -11,13 +11,8 @@ void NoFall::OnMoveCycle(DefaultMovementController* dmc, Vector3& offset) {
 }
 
 bool NoFall::CanExecute() {
-	if (!Util::IsValidPtr(Util::getLocalPlayer()))
-		return false;
-	if (!Util::IsValidPtr(Util::GetMovementController()))
-		return false;
-	return true;
+	return Util::isFullyInitialized();
 }
-
 
 
 void NoFall::Initialize() {
