+++ Hytale/Hooks/HookedMethods/DrawScene.cpp
@@ -1,0 +1,14 @@
+#include "../Hooks.h"
+#include "Events/EventRegister.h"
+
+void __fastcall Hooks::hkDrawScene(GameInstance* instance) {
+    Hooks::oDrawScene(instance);
+    if (!initialized)
+        return;
+
+    Renderer3D renderer3D;
+    EventRegister::Render3DEvent.Invoke(renderer3D);
+    renderer3D.Render();
+
+    fboRenderer->draw();
+}
