+++ Hytale/Hooks/HookedMethods/OnUserInput.cpp
@@ -1,0 +1,28 @@
+#include "../Hooks.h"
+
+void __fastcall Hooks::hkOnUserInput(uint64_t instance, SDL_Event event) {
+    Hooks::oOnUserInput(instance, event);
+
+    if (event.type != SDL_KEYDOWN && event.type != SDL_KEYUP)
+        return;
+
+    SDL_Scancode key{ event.key.scancode };
+    bool shouldCallOriginal = true;
+
+    if (event.type == SDL_KEYDOWN) {
+        if (event.key.repeat)
+            return;
+
+        InputSystem::inputMutex.lock();
+        InputSystem::keysPressed.insert(key);
+        InputSystem::keysHeld.insert(key);
+        InputSystem::keysUnheld.erase(key);
+        InputSystem::inputMutex.unlock();
+    } else if (event.type == SDL_KEYUP) {
+        InputSystem::inputMutex.lock();
+        InputSystem::keysHeld.erase(key);
+        InputSystem::keysUnheld.insert(key);
+        InputSystem::keysDepressed.insert(key);
+        InputSystem::inputMutex.unlock();
+    }
+}
