+++ Hytale/Hytale.vcxproj
@@ -174,6 +174,15 @@
     <ClCompile Include="Features\ActualFeatures\Nametags.cpp" />
     <ClCompile Include="Features\ActualFeatures\NoFall.cpp" />
     <ClCompile Include="Features\ActualFeatures\Speed.cpp" />
+    <ClCompile Include="Hooks\HookedMethods\DoMoveCycle.cpp" />
+    <ClCompile Include="Hooks\HookedMethods\DrawEntityCharactersAndItems.cpp" />
+    <ClCompile Include="Hooks\HookedMethods\DrawScene.cpp" />
+    <ClCompile Include="Hooks\HookedMethods\FrameIterator_IsValid.cpp" />
+    <ClCompile Include="Hooks\HookedMethods\HandleScreenShotting.cpp" />
+    <ClCompile Include="Hooks\HookedMethods\OnChat.cpp" />
+    <ClCompile Include="Hooks\HookedMethods\OnUserInput.cpp" />
+    <ClCompile Include="Hooks\HookedMethods\WeatherUpdate.cpp" />
+    <ClCompile Include="Hooks\HookedMethods\WglSwapBuffers.cpp" />
     <ClCompile Include="Hooks\Hooks.cpp" />
     <ClCompile Include="Math\Vector4.h" />
     <ClCompile Include="Menu\Children\ColorButton.cpp" />
