+++ Hytale/Util/Util.h
@@ -53,6 +53,7 @@ namespace Util {
     uint64_t BuildTicksFromHours(float hours);
     const char* GetKeyName(SDL_Scancode key);
     bool ShouldInteractWithGame();
+    bool isFullyInitialized();
     void allocate_console();
     void free_console();
     void log(const char* fmt, ...);
